refactor:update file name for onboarding style
