+++ app/onboardingStyles.ts
